Updated Boss & UI (PLEASE READ DESCRIPTION!)
Module Boss added and made it a bit functional, it still has to be polished (position, attacks...) but please, do not add attacks like the mouth one, I think is unnecessary at requires A FCKING LOT of time, focus first in finishing this shit pls

Changed bullets life to 1300 so they could reach the camera limit and not disappearing before

NOW THIS IS IMPORTANT:

Now each player has 3 lives. I have added to the UI a counter for each of them :), IF YOU SEE THE ORIGINAL GAME, aside of each life counter there's a little ship image, @adrianfr99 will have to put this, everything is set up so you can put it.

I have putted the "X" of both players in the exact site so that image can fit perfectly :)

And by the way, I know that the Player 2 UI is flipped, I cannot change that, it will be like this in this deliver, let it be :)

Kisses <3